requirements.txt
Django==2.2.5
django-cors-headers==3.1.1
djangorestframework==3.10.3
mysql-connector-python==8.0.17
mysqlclient==1.4.4
protobuf==3.9.2
PyMySQL==0.9.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
